feat(tck): add reset method
